move gracefulItems to agility and add agility cape to it
